// 1-front/src/app/_services/generated-api/model/decisionFiltrageQpcRowDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DecisionFiltrageQpcRowDTO { 
    id: number;
    ordreJuridictionnel: DecisionFiltrageQpcRowDTO.OrdreJuridictionnelEnum;
    juridiction: DecisionFiltrageQpcRowDTO.JuridictionEnum;
    niveauFiltrage: DecisionFiltrageQpcRowDTO.NiveauFiltrageEnum;
    niveauCompetence: string;
    dateFiltrage: string;
    formationJugement: string;
    numeroDecision: string;
    decisionRenvoi: string;
    decisionNonRenvoi: string;
}
export namespace DecisionFiltrageQpcRowDTO {
    export type OrdreJuridictionnelEnum = 'JUDICIAIRE' | 'ADMINISTRATIF';
    export const OrdreJuridictionnelEnum = {
        Judiciaire: 'JUDICIAIRE' as OrdreJuridictionnelEnum,
        Administratif: 'ADMINISTRATIF' as OrdreJuridictionnelEnum
    };
    export type JuridictionEnum = 'CONSEIL_ETAT' | 'COUR_DE_CASSATION';
    export const JuridictionEnum = {
        ConseilEtat: 'CONSEIL_ETAT' as JuridictionEnum,
        CourDeCassation: 'COUR_DE_CASSATION' as JuridictionEnum
    };
    export type NiveauFiltrageEnum = 'NONE' | 'PREMIER_ET_DERNIER' | 'DEUXIEME_FILTRAGE';
    export const NiveauFiltrageEnum = {
        None: 'NONE' as NiveauFiltrageEnum,
        PremierEtDernier: 'PREMIER_ET_DERNIER' as NiveauFiltrageEnum,
        DeuxiemeFiltrage: 'DEUXIEME_FILTRAGE' as NiveauFiltrageEnum
    };
}