// 1-front/src/app/_services/generated-api/model/pageDTODecisionFiltrageQpcRowDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DecisionFiltrageQpcRowDTO } from './decisionFiltrageQpcRowDTO';


export interface PageDTODecisionFiltrageQpcRowDTO { 
    totalElements: number;
    page: number;
    size: number;
    content: Array<DecisionFiltrageQpcRowDTO>;
}